refactor: clean up write decorator tests
